// client/src/lib/cloudinary/types.ts
export interface CloudinaryUploadResponse {
    public_id: string;
    version: number;
    signature: string;
    width: number;
    height: number;
    format: string;
    resource_type: string;
    created_at: string;
    tags: string[];
    bytes: number;
    type: string;
    etag: string;
    placeholder: boolean;
    url: string;
    secure_url: string;
    folder: string;
    original_filename: string;
}

export interface CloudinaryConfig {
    cloudName: string;
    uploadPreset: string;
    folder?: string;
}